requirements.txt
numpy==1.16.2
#tensorflow==1.14.0
scikit-learn==0.20.3
scikit-image==0.14.2
imageio==2.5.0
medpy==0.4.0
Pillow==6.0.0
scipy==1.2.1
pandas==0.24.2
tqdm==4.32.1
torch
torchvision